Fix reflection bug
This might seem like an odd thing to need to fix, but bear with me as I explain.

In certain environments, such as Fabric, the class loaders actually change through the loading process. In the case of reflection, it seems that internally google's ClassPath stuff actually changes the class loader that the class loads from. This causes some serious identity problems, such as the class not being able to recognize its parents (in this case, the "type" we are searching for). Although this should be impossible, the VM falls apart here, so in order to circumvent this, we forcefully load the class through the shared loader. This should never fail as the class always exists within this loader (ClassPath searches it).