deprecate JavaExecutable
it seems getResourceAsStream on fabric works as intended now. maybe i never did it correctly the first time, i wouldn't know anymore. However, this means JavaExecutable is no longer a necessity. I'll keep it here for compatibility sake, maybe we can find use for it in a rewrite, but otherwise it is no longer necessary as it once was.