bring back java executable
fabric loader is annoying. tl;dr at load time the class loader either have no permission to access the files, or simply cannot given how jars are loaded into the mc environment. as a result, while getResource works for development, it will not work outside of it. as a result, our old hacky alternative to access the jar data via ZipFileStore is still necessary for outside of the dev environment